Se separan ruta y servicio de geonode
